chore: Base of Isshub project
Abstract
========

The base of the project, including setup files, README, LICENSE and
gitignore.

Motivation
==========

Everything needs a starting point.

Rationale
=========

Readme
------

Describe the projects, the why and the how.

Actually only includes the main abstract parts of of the README, in
restructured text.

Format
''''''

The choice of this format was made to ease the integration with sphinx
to build the documentation.

Sections
''''''''

The README includes these sections:

- Abstract: To explain in a few words what this repository is.
- Motivation: To explain why a new version of isshub is needed.
- Rationale: The choices made before starting.

setup
-----

Base setup for the project, with version "0" (nothing inside for now).

setup.cfg
'''''''''

I chose the `setup.cfg` way of doing things because configuring a
project must be, in my opinion, declarative, and not code.

This will also allow other tools to have their configuration in the same
place.

License
-------

I'm not sure yet so for now it will be MIT. So this commit also includes
the LICENSE file, and defines it in `setup.cfg`

Makefile
--------

To help usage of the different tools that will be added, `make` is a
common and useful tool. So here it's base version with a few commands:
- help: Show the list of commands
- install: Install the project in the current environment, with its
   dependencies
- dev: Install the project in the current environment, with its
   dependencies, including the ones needed in a development environment
- dev-upgrade: Upgrade all default+dev dependencies defined in
   setup.cfg
- dist: Build the package