style: Add black and isort to ensure code quality
Abstract
========

Install and configure black and isort to force the code to be well
written.

Motivation
==========

We previously installed pylint and flake8 to ensure the code is well
written, in a clean and common style used among the whole project.

But coding, waiting for the CI just to correct styling is overwhelming.

Rationale
=========

Black is the uncompromising Python code formatter. So every code will be
formatted the same way, no matter what.

isort is a Python utility / library to sort imports alphabetically, and
automatically separated into sections.

5 make commands are added:

- `make isort`: Run the isort tool and update files that need to
- `make black`: Run the black tool and update files that need to
- `make check-isort`: Run the isort tool in check mode only (won't
   modify files)
- `make check-black`: Run the black tool in check mode only (won't
   modify files)
- `make pretty`: Run all code beautifiers (isort, black)

And `make lint` is updated to add `check-isort` and `check-black`