style(git): Force format of git commit messages
Abstract
========

Enforce a very specific style for commit messages.

Motivation
==========

Commits are as much important as code. They hold whole history of the
project and commits can be used to know why things were done a certain
way.

So to keep the information in the commits, we enforce a style for commit
messages.

Rationale
=========

`conventional commits <https://www.conventionalcommits.org>`_ is great
but lacks style in the "description" part. So we use it only for the
"subject" part

In the description we enforce restructed text and sections like for the
python PEPs.

A new `make` command is added to ensure the last commit is valid,
`make check-commit`, and the `make checks` command is updated to include
this one too.

To check for other commits that the last one, or text that is not
already used, use `ci/check_commit_message.py`. Call it with `-h` to
show how to use it.