ci: Configuration for circle-ci
Abstract
========

This commits adds the configuration file to use `circle-ci
<https://circleci.com>`_

Motivation
==========

Running all checks via make locally is great but having this done
externally and automatically is better. For external pull-requests but
not only.

And it allows to run jobs that are long to run locally and where a
failure is rare, but that are run "just in case", like building the
project and running tests on the build.

Rationale
=========

I chose to use circle-ci because the ease of how we can define
dependencies.

All jobs simply call `make` commands (after a bit of preparation)