tests: Setup testing tools with pytest
Abstract
========

Use pytest (with some sugar) for tests.

Motivation
==========

Code is not known to be working if it's not tested. So tests are
mandatory.

Rationale
=========

`pytest` is better suited to not django-only projects, which is the case
here. Also with pytest we avoid inheritance hell in tests: a test must
be fully understandable while reading at its code.

We also install:
- pytest-cov: to enable test coverage
- pytest-sugar: for a "beautiful" output of tests.

A new `make` command is added to run the tests::

    make tests

To run a specific tests::

    pytest python.path.to.tests::function-to-test

This commit includes a `test_testing.py` file to ensure the CI catches
the tests.