style: Add mypy for better code quality
Abstract
========

Install and configure black and isort to force the code to be well
written.

Motivation
==========

Having documentation in every method is great but nothing ensure that
the it is up to date. Having a tool to check that at least the
documented types of attributes and return values are correct and
correctly used from one function to another is great.
We previously installed pylint and flake8 to ensure t

Rationale
=========

Python provide "types annotation". And mypy will check the project that
they are correctly used.

As it can be cumbersome to use, we choose to not use them in tests.

1 make command is added:

- `make mypy`: Run the mypy tool

And `make lint` is updated to add `mypy`