ci: Make the CI do all checks for every commits
Abstract
========

Make the CI run all checks (lint, tests, commit message) for all commits
in a branch.

Motivation
==========

The state of the project after every commit must be a valid state: tests
must pass. It is absolutely necessary when you want to find the source
of a bug via `git-bisect`.

And enforcing the validity of the commit message must of course be done
for every commit in a branch.

Rationale
=========

This is done by a new job in the circl-ci configuration that will call a
bash script in `ci/check-every-commit.sh`.

This script will run `make lint`, `make tests` and `make check-commit`
for every new commit starting from the branch starting point.

It will be done only once for a commit and the result will be stored in
a GIST, and a "status" will be added, for each command, on the commit on
the Github side.