docs(rtd): Add configuration for readthedocs.org
Abstract
========

Add the `.readthedocs.yml` file for readthedocs configuration

Motivation
==========

Having the documentation able to be built is ok, having a place to
display it is better.

Rationale
=========

The master in this domain, especially for python projects, is
readthedocs.org

So we configure it to build the project installing the package with
`docs` dependencies.