docs(git): Add git commits to documentation
Abstract
========

This commit adds all the commits into the generated documentation.

Motivation
==========

As stated previously, "commits are as much important as code". And now
that we have all commits following the same template with informative
descriptions, we can have them all in the documentation to have the
history of the project documented.

Rationale
=========

The choice was made to generate RST from commits infos, instead of
directly generated documentation page in html, or creating "rst python
code".

Se the generated files are readable even outside of the html build.

The RST is constructed via simple templates, and GIT is read via
pydriller + gitpython.

The generated files includes:
- all branches
- all tags
- all commits with diffs + links to parent commits and children
- all files with their history

The GIT documentation is included in the `internal` parts of the full
documentation.