tests: Remove pure testing tests
Abstract
========

Remove `test_testing.*` files.

Motivation
==========

Now that we know that tests are working, we can remove the ones that
purely existed to check if the CI did its job right.

Rationale
=========

In the makefile, we ignore "error 5" from pytest, error simply meaning
that there is no test to run.