feat(repository): Add 1st domain context (core) and entity (Repository)
Abstract
========

Add the `core` domain context with the `Repository` entity

Motivation
==========

Domain Driven Developement is a great way to have separation of
concerns.

Bounded contexts are used to separate groups of things that work
together but independant of other contexts.

Entities are objects that have a distinct identity.

As we need to start somewhere, this commit introduce DDD with its first
context and first entity, and test it using BDD.

Rationale
=========

As the code repositories are at the root of Isshub, this commit
introduce the `Repository` as its first entity, in a `core` context that
will also holds issues, code requests, etc...

An entity must have a specific set of fields and this is defined and
tested using BDD, in a `describe.feature` file.

We make pylint and flake8 ignore test files because:
- no need for documentation as the BDD decorators include what is done
in a function
- fixture usage is "magic" so it generates some warnings
- we still enforce black and isort so it's enough for having good code

And finally we updated how tests are run in CI for built packages, now
that tests are not in a main `tests` directory, but near the things they
test. we also do a better exclusion of tests in the final build and
added LICENSE and README.rst to it.