feat(namespace): Add Namespace entity in core domain context
Abstract
========

Add the `Namespace` entity in the `core` domain context. And add testing
utilities to help testing things that look the same.

Motivation
==========

On Github, repositories are stored under the name of their owner, which
can be users or organizations.

On Gitlab, repositories are stored under namespaces, which can be their
owner if it's a user, or a group (which can also have a parent group,
etc)

So the concept of "namespace" allows to represent all of these.

Rationale
=========

The type of the `namespace` field of the `Repository` entity is changed
from `str` to `Namespace`, and is still mandatory.

The `Namespace` entity also has a `namespace` field, which is an
optional `Namespace` (if None, it won't have any parent).

We had to do a "trick" to handle a self-reference in type for this
`namespace` field. We created a `_Namespace` class, defining everything,
and a `Namespace` one, inheriting from it, simply setting the
`namespace` attribute as an optional `Namespace`.

Many testing helpers have beed added to
`isshub.domain.utils.testing.validation`, because the tests for the
`Namespace` entity needed nearly the same code as the tests for the
`Repository` entity.

To have clear pytest error messages, we told pytest to handle the assert
in this file the same way as in test files, via
`pytest.register_assert_rewrite`.