feat(namespace): Add Namespace entity a `kind` field
Abstract
========

Adds a `kind` field to the `Namespace` entity of the `core` domain
context.

Motivation
==========

Even if the name of a namespace is the most important thing, it is
useful to know which kind of namespace it is.

Rationale
=========

We use an python `Enum` as it's the simplest way to represent a list of
choices.

We add a test dependency, `fake-enum` to easily add the `kind` field to
the `Namespace` factory.