style(git): Add list of available types in git commit message
Abstract
========

Add a line in `.gitmessage` with the list of available types.

Motivation
==========

Having to remember the exact types is cumbersome, having it right at
disposal without having to look at documentation is better.

Rationale
=========

The line must be removed for the commit to be valid, so a note at the
end tells it to the user.