docs(fix): Fix usage of `pydriller` in `git_to_sphinx`
Abstract
========

Fix code to get the "path to repo" from `pydriller`

Motivation
==========

It was not seen in the job running `make doc` because warnings are
ignored (to be changed in a future commit).

Rationale
=========

They store a list now so we need to get the first item from this list.