ci(docs): Make CI job `build_doc` fail if a warning occurs
Abstract
========

Add a new make command, `doc-strict`, that will fail if a warning is
generated during the build, and use this command in the `build_doc` CI
job

Motivation
==========

As seen in the two previous commits, we weren't able to see two failures
during the build of the documentation:
- badly formatted README file
- error when running `git_to_sphinx`

It was only visible when looking at the documentation.

Rationale
=========

`sphinx-build` accepts a `-W` argument that treats warnings as errors.

It's not possible to use it with the `make` commands of the docs
MakeFile, so we had to put the whole `sphinx-build` command in the `make
doc-strict` command.