docs(git): Show source of renamed files
Abstract
========

When a file is renamed, show the last source on the pages of the old and
new paths, by following the modifications tree to look for the last
modification with a source code.

Motivation
==========

When we renamed the `core` domain context to `code_repository`, the
source code of all included files disapeared from the "git" part of the
documentation.

Rationale
=========

A new function `find_last_source_code` is added in `git_to_sphinx.py`
that will be executed recursively if the source code of file is not
found, by checking in the parent modification if any, or the one of the
old path in case of rename.