docs(git): Only update remote branches if asked
Abstract
========

In `git_to_sphinx.py`, only fetch remote branches to update them locally
if the `GIT_TO_SPHINX_UPDATE_BRANCHES` environment variable is set.

Force it to be set for CI when running `make doc-string`, and for
readthedocs.

Motivation
==========

When working on a branch, with some commits applied but not pushed to
the origin, running `git_to_sphinx` (or by just doing `make docs`)
reseted the branches to their remote head. so newly added commits where
lost (not really, thanks to reflog)

Rationale
=========

I chose to use a environment variable to make this choice really a
choice, not an accident.