docs(examples): Add examples in functions docstring
Abstract
========

Add some docstring example to many functions

Motivation
==========

Having functions with docstring explaining what they do, what are
they arguments and their purpose and what they return is great, but
having an example on how to use it with its behavior is a great help.

Rationale
=========

We use the `Examples` section of the docstring.
These examples are tested by pytest (via pytest and the
`doctest-modules` argument) so we are sure the examples are working.