docs(source): Reduce white space in toc tree
Abstract
========

In the toc tree of the "Packages" section, reduce the space above and
below the "sub packages".

Also raise the depth of displayed sub packages from 4 to 6.

Motivation
==========

Readability in documentation is important and there was too much
whitespace, and not enough depth.

Rationale
=========

For the white space, we simply removed the margin in sub ul elements.

For the toc tree depth, we passed it when calling `apidoc`