build(python): Update minimal python version to 3.8
Abstract
========

We change the python version to use in readthedocs and circle ci
configuration files, and in setup.cfg to make sure users will use the
correct version.

Motivation
==========

At the time this project was created, python 3.8 existed but was not
available on external tools like readthedocs and circleci.

They are now, so we can update the python version everywhere.

Rationale
=========

N/A