build(make): Remove the `dev-upgrade` make command (use `make dev`)
Abstract
========

Remove the `dev-upgrade` command from the `Makefile`, and update the
`dev` command to work for installing or upgrading.

Motivation
==========

The `dev-upgrade` command could be simplified from

`pip install --upgrade `python -c 'import setuptools...`

to exactly the same content as the `dev` one, except for the `--upgrade`
part.

Rationale
=========

The `dev-upgrade` command was removed and the `--upgrade` argument was
added in the `dev` command.