ci: Run full workflow on `develop` every day
Abstract
========

Add a new workflow in circle-ci configuration to run the full workflow
on the `develop` branch every night.

Motivation
==========

To avoid being surprised after a long day without touching the project
when dependencies introduce breaking changes, it's useful to be alerted
as soon as possible. So having a nightly job running on the main
development branch, `develop` sounds like a good way to do it.

Rationale
=========

To avoid duplicating the list of jobs, we used the "reference" feature
of yaml to define the list of jobs to run with their dependencies only
once, using this reference in our two workflow:

- `full_workflow_on_push`, to be run on each push, renamed from `isshub`
- `full_workflow_nightly`, to be run every day at 3:21 am on `develop`