fix(namespace): Namespaces relationships should not create a loop
Abstract
========

Ensure that when settings a parent namespace to a namespace, we don't
create a loop, a loop being having a namespace descendant be the parent
of one of its ascendant.

Motivation
==========

It's not possible to have a namespace belonging to a namespace that is
one of its children namespace, or even itself.

Rationale
=========

This is done via a method decorated via `field_validator`, with a
simple recursive loop.