fix(entities): Entities `id` field are frozen once set
Abstract
========

Make the `id` field of all entities (via `BaseModelWithId`) frozen, ie
that cannot be updated once set at init time, else a
`FrozenAttributeError` exception is raised

Motivation
==========

An "id" identifies an object, so there is no point to make it editable.

Rationale
=========

This is done by using the `on_setattr` argument to `attr.ib` with the
`frozen` setter.

There is a main difference with other fields validators because this
is done at the moment the field is set, and not later while validating
it, because at validation time, the field is already set and there is no
clean way to get the old value.