docs(bdd): Add BDD scenarios to documentation
Abstract
========

Add a new "Specifications" entries at the root of the documentation,
containing all features and scenarios.

Motivation
==========

We use BDD as specification, so it's important to have the features and
scenarios in the documentation.

Rationale
=========

We use the `sphinx-gherkindoc` package to do the heavy work of finding
the `.feature` files and rendering the documentation.