style(mypy): Remove most "type: ignore" pragmas
Abstract
========

Be more precise about typing to be able to remove many or our pragmas
telling mypy to ignore some parts of the code, especially in the
`isshub.domain.utils.entity` module.

Motivation
==========

Simply calling `Repository(name=name, namespace=namespace)` made mypy
fail. So it was necessary to let mypy understand our layer above the
`attr` package. Also, having all these `# type: ignore` pragmas was
not really what we wanted, it was temporary until having time to fix it.

Rationale
=========

- Add a mypy plugin to let mypy know about our own functions to define
  `attr` classes and fields
- Add typing in the `isshub.domain.utils.entity` to be able to remove
  all our `# type: ignore` pragmas when defining classes and fields

During this last step, we had to change the `field_validator` decorator
from a function to a class because of an error from mypy that kept
telling us that the functions decorated by this decorated were made
untyped because the decorator was untyped... but it was not.