style(entity): Change the world "model" by "entity"
Abstract
========

Rename `BaseModel` and `BaseModelWithId` to `BaseEntity` and
`BaseEntityWithId`, and replace the word "model" by "entity" everywhere,
in comments and function names.

Motivation
==========

To keep things consistent, it's better to have one and only one word to
name a thing, and we started with entity, and it's the name used in the
DDD word.

Rationale
=========

N/A