fix(entity): `id` changed from `int` to `uuid4`, renamed to `identifier`
Abstract
========

The `id` field of entities is renamed to `identifier` and the type is
changed from `int`, validated as positive integer, to `uuid.UUID`,
validated to be a UUID version 4.

Motivation
==========

The `id` term is more a "technical" term, that is a shortened version
of the word `identifier`. And it cannot be used as an argument without
shadowing the `id` python builtin. So it's better to use the full word.

And being an integer means that checks have to be done for uniqueness.
UUIDs exists exactly to solve this.

Rationale
=========

N/A