feat(repository): Add domain repositories
Abstract
========

Add base repositories to store entities. With in-memory ones for
entities in the "code-repository" context (`Repository` and `Namespace`)

Motivation
==========

It's the next step in DDD: now that we have some entities, we need to
store them, and this is done via repositories.
For now, simply adding base repositories and in-memory ones is enough.
To save entities in a permanent way, we'll introduce later other
repositories via the Django ORM.

Rationale
=========

N/A